requirements.txt
python-dotenv
langchain
langchain-groq